refactor: wrap the components with fallback directory
