+++ app/(fallback)/group/detail/[id].tsx
