+++ app/(fallback)/index.tsx
@@ -1,0 +1,14 @@
+import { Link } from 'expo-router';
+import { View } from 'react-native';
+
+export default function Index() {
+  return (
+    <View>
+      <Link href="/login">Login</Link>
+      <Link href="/chat">Chat</Link>
+      <Link href="/bounty">Bounty</Link>
+      <Link href="/group">Group</Link>
+      <Link href="/group/create">Group Create</Link>
+    </View>
+  );
+}
